"Allocate" module and page deleting in destructor in progress

+++ ObjectAllocator-files/driver.cpp
@@ -22,8 +22,8 @@ int SHOW_EXCEPTIONS = 0;
 struct Student {
     int Age;
     float GPA;
-    long Year;
-    long ID;
+    long long Year;
+    long long ID;
 };
 
 struct Employee {
